Refactor to one module
