+++ scripts/run_finite_state_machine_visu.py
@@ -1,4 +1,6 @@
-from finite_state_machine import traffic_light_state_machine_with_visu
+from behavior_generation_lecture_python.finite_state_machine import (
+    traffic_light_state_machine_with_visu,
+)
 
 if __name__ == "__main__":
     tl = traffic_light_state_machine_with_visu.TrafficLightStateMachine()
