+++ src/behavior_generation_lecture_python/graph_search/a_star.py
@@ -1,5 +1,6 @@
-import matplotlib.pyplot as plt
 import math
+
+import matplotlib.pyplot as plt
 
 
 class GraphNode:
