+++ src/behavior_generation_lecture_python/lateral_control_riccati/riccati_controller.py
