+++ src/behavior_generation_lecture_python/utils/generate_reference_curve.py
@@ -1,6 +1,6 @@
+import matplotlib.pyplot as plt
 import numpy as np
 from scipy import interpolate
-import matplotlib.pyplot as plt
 
 
 def pick_points_from_plot():
