+++ src/behavior_generation_lecture_python/vehicle_models/dynamic_one_track_model.py
@@ -1,7 +1,10 @@
 import numpy as np
 from scipy.integrate import odeint
 
-from vehicle_models.vehicle_parameters import VehicleParameters, DEFAULT_VEHICLE_PARAMS
+from behavior_generation_lecture_python.vehicle_models.vehicle_parameters import (
+    DEFAULT_VEHICLE_PARAMS,
+    VehicleParameters,
+)
 
 
 class DynamicOneTrackModel:
